decimal changelist

+++ app/src/main/java/com/wang17/myphone/model/StockInfo.java
@@ -1,13 +1,15 @@
 package com.wang17.myphone.model;
 
 
+import java.math.BigDecimal;
+
 public class StockInfo {
     public String time;
     public String name;
     public String code;
-    public double cost;
-    public double price;
-    public double increase;
+    public BigDecimal cost;
+    public BigDecimal price;
+    public BigDecimal increase;
     public String exchange;
     public int amount;
     public int type;
